+++ autorest/codegen/templates/init.py.jinja2
@@ -12,7 +12,7 @@ __all__ = ['{{ code_model.class_name }}']
 {% if not async_mode %}
 
 try:
-    from .patch import patch_sdk
+    from ._patch import patch_sdk
     patch_sdk()
 except ImportError:
     pass
