+++ test/azure/Expected/AcceptanceTests/HeadExceptions/headexceptions/__init__.py
@@ -13,7 +13,7 @@ __version__ = VERSION
 __all__ = ['AutoRestHeadExceptionTestService']
 
 try:
-    from .patch import patch_sdk
+    from ._patch import patch_sdk
     patch_sdk()
 except ImportError:
     pass
