+++ test/multiapi/Expected/AcceptanceTests/Multiapi/multiapi/v1/__init__.py
@@ -10,7 +10,7 @@ from ._multiapi_service_client import MultiapiServiceClient
 __all__ = ['MultiapiServiceClient']
 
 try:
-    from .patch import patch_sdk
+    from ._patch import patch_sdk
     patch_sdk()
 except ImportError:
     pass
