+++ test/vanilla/Expected/AcceptanceTests/AdditionalProperties/additionalproperties/__init__.py
@@ -13,7 +13,7 @@ __version__ = VERSION
 __all__ = ['AdditionalPropertiesClient']
 
 try:
-    from .patch import patch_sdk
+    from ._patch import patch_sdk
     patch_sdk()
 except ImportError:
     pass
